requirements.txt
Pillow>=10.0.0
rich>=13.0.0
click>=8.0.0
pytest>=7.0.0
